requirements.txt
streamlit
python-dotenv
google-generativeai
python-Levenshtein
fpdf2         
openpyxl      
pandas  
requests